+++ hydra-cluster/src/Hydra/Cluster/Scenarios.hs
@@ -2364,10 +2364,7 @@ waitsForChainInSyncAndSecure tracer workDir backend hydraScriptsTxId = do
         send n3 $ input "NewTx" ["transaction" .= tx]
 
         waitMatch 5 n3 $ \v -> do
-          guard $ v ^? key "tag" == Just "RejectedInput"
-          guard $ case v ^? key "reason" . _String of
-            Just reason -> "chain out of sync" `T.isPrefixOf` reason
-            Nothing -> False
+          guard $ v ^? key "tag" == Just "RejectedInputBecauseUnsynced"
 
         -- Carol API notifies the node is back on sync with the chain
         -- note this is tuned based on how long it takes to sync